+++ include/Rating.h
