+++ src/GridLayout.cpp
@@ -1,4 +1,4 @@
-#include "includes/GridLayout.h"
+#include "GridLayout.h"
 
 #include <stdexcept> // used for std::out_of_range
 #include <iostream>
