+++ src/Menu.cpp
@@ -1,4 +1,4 @@
-#include "includes/Menu.h"
+#include "Menu.h"
 
 Menu::Menu()
 {
