+++ src/Option.cpp
@@ -1,4 +1,4 @@
-#include "includes/Option.h"
+#include "Option.h"
 
 float Option::getVolume()
 {
